Use worker pattern
